+++ src/lib/searchDocument.ts
@@ -1,9 +1,10 @@
 import { search } from '@ng2react/core';
-import * as vscode from 'vscode';
-import { assertJs } from './validation';
+import * as fs from 'fs';
 import { startCase } from 'lodash';
 import * as path from 'path';
-import * as fs from 'fs';
+import * as vscode from 'vscode';
+import { getSourceRoot } from '../Config';
+import { assertJs } from './validation';
 export default function searchDocument(document: vscode.TextDocument) {
     assertJs(document);
     return search(document.getText(), { file: document.fileName });
@@ -21,7 +22,7 @@ export default function searchDocument(document: vscode.TextDocument) {
 export function findExistingConversions(componentFile: vscode.Uri, componentName: string) {
     const componentFileName = componentFile.path.split('/').pop()!;
     const ext = componentFileName.split('.').pop()! === 'ts' ? 'tsx' : 'jsx';
-    const componentFileDir = componentFile.path.replace(componentFileName, '');
+    const componentFileDir = getReactFolder(componentFile);
     const reactElementName = startCase(componentName).replace(/ /g, '');
     const convertedComponentFileName = `${reactElementName}.${ext}`;
     const convertedComponentFilePath = path.join(componentFileDir, convertedComponentFileName);
@@ -30,4 +31,21 @@ export function findExistingConversions(componentFile: vscode.Uri, componentName
         return [];
     }
     return [{ name: reactElementName, uri: vscode.Uri.file(convertedComponentFilePath), type: 'react' as const }];
+}
+
+function getReactFolder(angularFile: vscode.Uri) {
+    const angularRoot = getSourceRoot('angular');
+    const reactRoot = getSourceRoot('react');
+
+    const angularFileDir = path.dirname(angularFile.path);
+
+    if (!(angularRoot && reactRoot) || angularRoot === reactRoot) {
+        return angularFileDir;
+    }
+
+    if (!angularFileDir.includes(angularRoot)) {
+        return angularFileDir;
+    }
+
+    return angularFileDir.replace(angularRoot, reactRoot);
 }